build: introduce nonshared repository configuration
Introduce a new menuconfig option VERSION_NONSHARED_REPO, which
allows users to set a separate URL specifically for target specific
(nonshared) packages repository.

We use a new %N placeholder to represent this new repository URL
for the core and kmods feeds. The base feed and other enabled feeds
continue to use the standard VERSION_REPO URL.

This flexibility is useful for scenarios where target specific
binaries and kernel modules are hosted on a different mirror or path.

Signed-off-by: Ethan Zuo <yuxuan.zuo@outlook.com>